requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
